nsscash: main_test: fix Last-Modified/If-Modified-Since headers
http.TimeFormat requires a time in UTC for a proper Last-Modified
header. Previously, the time was in CEST (my local timezone) which
caused all Last-Modified headers to have an offset of two hours. This
also hid a bug in the If-Modified-Since header handling: The last change
time must be truncated to seconds as the If-Modified-Since header has
only a second precision.

These bugs did not affect the current test cases. However, it affects
the test case which will be added in the next commit.