nsscash: main_test: test new server response which causes an update
The existing tests only tested that a modified local version of the file
causes a download from the server. However, the modified local version
is caught by the checksum in the state file and does not verify that an
update from the server is properly applied. Thus, add this new test.